data-ingestion/requirements.txt
apache-airflow-providers-google
apache-airflow-providers-apache-spark
kaggle
findspark
